adjust IntroMusic2 Intro Time. sounds a lot better now

+++ Assets/Script/GameManager.cs
@@ -124,7 +124,7 @@ public class GameManager : MonoBehaviour {
 		}
 
 		if (! gameStarted) {
-			startmsgs.transform.FindChild("Countdown").GetComponent<Text>().text = ((int) introTime - (Time.time - startTime)).ToString ();
+			startmsgs.transform.FindChild("Countdown").GetComponent<Text>().text = ( introTime - (Time.time - startTime)).ToString ();
 			Player1.GetComponent<Attackable> ().health = 100;
 			if (Input.GetKeyDown(KeyCode.Escape)) {
 				startGame ();
